MAT-9203: Refactor testcases/components/common
