+++ src/components/editMeasure/testCases/components/QDM/componentDataType/ComponentType.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import "twin.macro";
 import "styled-components/macro";
-import CodeInput from "../codeInput/CodeInput";
+import CodeInput from "../../QDM/codeInput/CodeInput";
 import { Select } from "@madie/madie-design-system/dist/react";
 import { ValueSet, DataElement, Component } from "cqm-models";
 import { MenuItem } from "@mui/material";
