+++ src/components/editMeasure/testCases/components/editTestCase/qiCore/LeftPanel/ElementsTab/builder/element/types/CodeableConceptComponent.tsx
@@ -3,7 +3,6 @@ import "twin.macro";
 import "styled-components/macro";
 import { Coding } from "fhir/r4";
 import CodingComponent from "./CodingComponent";
-import AddElementButton from "../../../../../../../common/AddElementButton";
 
 const CodeableConceptComponent = ({
   canEdit,
