+++ src/components/editMeasure/testCases/components/editTestCase/qiCore/LeftPanel/ElementsTab/builder/element/types/InstantComponent.tsx
@@ -3,7 +3,7 @@ import { Instant } from "@madie/madie-design-system/dist/react";
 
 import dayjs from "dayjs";
 import { INSTANT_REGEX } from "../typesValidations/fhirR4Validations";
-import AddElementButton from "../../../../../../../common/AddElementButton";
+import AddElementButton from "../../../../../../../common/UIOnlyModelAgnostic/AddElementButton";
 
 const isValidInstant = (instantString: string) => {
   if (!INSTANT_REGEX.test(instantString)) return false;
